requirements_dev.txt
apache-libcloud==1.2.1
netifaces
bumpversion==0.5.3
paramiko
wheel
watchdog==0.8.3
Sphinx==1.3.1
cryptography
PyYAML==3.11
requests
requests_ntlm
-e git+https://github.com/diyan/pywinrm.git#egg=pywinrm
colorlog
six
certifi
xmltodict

